requirements.txt
ansible==2.14.0
pywinrm>=0.4.3
molecule>=5.0.0
molecule-plugins[docker]>=23.5.0
pytest-testinfra>=8.1.0
ansible-lint>=6.14.0
yamllint>=1.29.0
pytest>=7.3.1
python-dotenv>=1.0.0 